Refactor expressions to share a common base node
