+++ pyomo/core/expr/expr_common.py
@@ -9,6 +9,11 @@
 #  This software is distributed under the 3-clause BSD License.
 #  ___________________________________________________________________________
 
+import enum
+
+from pyomo.common.backports import nullcontext
+from pyomo.common.deprecation import deprecated
+
 TO_STRING_VERBOSE=False
 
 _add = 1
@@ -63,3 +68,42 @@ _inv = 2
 _equiv = 3
 _xor = 4
 _impl = 5
+
+class OperatorAssociativity(enum.IntEnum):
+    """Enum for indicating the associativity of an operator.
+
+    LEFT_TO_RIGHT(1) if this operator is left-to-right associative or
+    RIGHT_TO_LEFT(-1) if it is right-to-left associative.  Any other
+    values will be interpreted as "not associative" (implying any
+    arguments that are at this operator's _precedence() will be enclosed
+    in parens).
+
+    """
+
+    RIGHT_TO_LEFT = -1
+    NON_ASSOCIATIVE = 0
+    LEFT_TO_RIGHT = 1
+
+@deprecated("""The clone counter has been removed and will always return 0.
+
+Beginning with Pyomo5 expressions, expression cloning (detangling) no
+longer occurs automatically within expression generation.  As a result,
+the 'clone counter' has lost its utility and is no longer supported.
+This context manager ill always report 0.""", version='TBD')
+class clone_counter(nullcontext):
+    """ Context manager for counting cloning events.
+
+    This context manager counts the number of times that the
+    :func:`clone_expression <pyomo.core.expr.current.clone_expression>`
+    function is executed.
+    """
+    _count = 0
+
+    def __init__(self):
+        super().__init__(enter_result=self)
+
+    @property
+    def count(self):
+        """A property that returns the clone count value.
+        """
+        return clone_counter._count
